refactor: relocate database types to src/shared/types/database.ts
Move all pure data interfaces from src/main/database/types.ts to
src/shared/types/database.ts so that src/shared/ no longer imports
from src/main/. The original file is replaced with a re-export shim
so all existing main-process imports remain unaffected.
